changes after review, not working yet
